tool/sumbot/requirements.txt
# API 框架
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5

# OpenAI
openai>=1.0.0

# 测试
pytest>=8.3.4
pytest-asyncio>=0.25.0

# 内容处理
markdown>=3.7.0
beautifulsoup4>=4.9.3
aiohttp>=3.8.0
python-docx>=1.0.1
PyMuPDF>=1.23.8
openpyxl>=3.1.2
python-pptx>=0.6.22

# 工具
python-dotenv>=1.0.0
httpx>=0.26.0
requests>=2.31.0
psutil>=5.9.7
aiofiles>=0.8.0 